Consolidate entity tests into single test_secret.py file
Combined test_endpoint.py, test_secret.py, and test_secret_binding.py
into a single test file since these entities are all part of the same
cohesive secrets/endpoints feature.

Organization:
- Secret Tests (4 tests) - Secret entity and SecretTag
- Endpoint Tests (6 tests) - Endpoint entity and EndpointTag
- SecretBinding Tests (5 tests) - Binding entity and resource types
- Wrapper Entity Tests (1 test) - SecretWithEndpointAndBinding

Benefits:
- All related tests in one place
- Easier to understand the full feature scope
- Reflects the tight coupling of these entities
- Cleaner test directory structure

All 16 tests passing.
